+++ src/main/java/lotto/domain/ticket/Lotto.java
@@ -1,11 +1,11 @@
-package lotto.domain.lotto;
+package lotto.domain.ticket;
 
-import static lotto.domain.lotto.LottoErrorMessage.DUPLICATION_ERROR;
-import static lotto.domain.lotto.LottoErrorMessage.INVALID_RANGE_ERROR;
-import static lotto.domain.lotto.LottoErrorMessage.SIZE_ERROR;
-import static lotto.domain.lotto.LottoNumberConfig.MAXIMUM_RANGE;
-import static lotto.domain.lotto.LottoNumberConfig.MINIMUM_RANGE;
-import static lotto.domain.lotto.LottoNumberConfig.NUMBER_COUNT;
+import static lotto.domain.ticket.LottoErrorMessage.DUPLICATION_ERROR;
+import static lotto.domain.ticket.LottoErrorMessage.INVALID_RANGE_ERROR;
+import static lotto.domain.ticket.LottoErrorMessage.SIZE_ERROR;
+import static lotto.domain.ticket.LottoNumberConfig.MAXIMUM_RANGE;
+import static lotto.domain.ticket.LottoNumberConfig.MINIMUM_RANGE;
+import static lotto.domain.ticket.LottoNumberConfig.NUMBER_COUNT;
 
 import java.util.ArrayList;
 import java.util.List;
